added order service

+++ Backend/Inventory-service/src/main/java/com/example/inventoryService/dto/InventoryResponse.java
@@ -15,8 +15,8 @@ public class InventoryResponse {
     private Long inventoryId;
     private String productId;
     private String sku;
-    private long totalQuantity;
-    private long reservedQuantity;
+    private Integer totalQuantity;
+    private Integer reservedQuantity;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
